refactor(server): clean up cf package: direct imports, f-string, docstrings
